Fix id variable type ( int )
git-tfs-id: [http://developmentvlan:8585/tfs/vanrise.collection]$/;C24313

+++ BusinessProcess/Code/Vanrise.Queueing.Data.SQL/QueueDataManager.cs
@@ -154,7 +154,7 @@ namespace Vanrise.Queueing.Data.SQL
         {
             return new QueueItemType
             {
-                Id = (long)reader["Id"],
+                Id = (int)reader["Id"],
                 ItemFQTN = reader["ItemFQTN"] as string,
                 CreatedTime = GetReaderValue<DateTime>(reader, "CreatedTime"),
                 DefaultQueueSettings = reader["DefaultQueueSettings"] as string,
